# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fly Delivery",
    "description": "This application offers a seamless and convenient way to satisfy your cravings with just a few clicks. Fly-Delivery connects you with a wide range of restaurants and cuisines, bringing delicious meals straight to your doorstep."
  },
  "servers": [
    {
      "url": "http://localhost:4444/"
    }
  ],
...
  "paths": {
...
    "/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Register new user",
        "description": "Register new user",
        "operationId": "regsterNewUser",
        "requestBody": {
          "description": "Register new user",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "exemple"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "exemple@email.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "*******"
                  },
                  "role": {
                    "type": "string",
                    "example": "Customer"
                  },
                  "phone": {
                    "type": "string",
                    "example": "00 000 000"
                  },
                  "adress": {
                    "type": "string",
                    "example": "example"
                  },
                  "avatar": {
                    "type": "string",
                    "example": "example"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request! Email already exist!"
          },
          "500": {
            "description": "Internal Server Error!"
          }
        }
      },
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "users"
        ],
        "summary": "Get the list of all users",
        "description": "Get the list of all users",
        "operationId": "ListUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": true,
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": true,
            "type": "integer"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Field to sort by",
            "required": true,
            "type": "string"
          },
          {
            "name": "createdAtBefore",
            "in": "query",
            "description": "Filter users created before this date (YYYY-MM-DD format)",
            "required": true,
            "type": "string"
          },
          {
            "name": "createdAtAfter",
            "in": "query",
            "description": "Filter users created after this date (YYYY-MM-DD format)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successfuly operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 1
                    },
                    "totalUsers": {
                      "type": "integer",
                      "example": 1
                    },
                    "users": {
                      "type": "array",
                      "example": [
                        {
                          "_id": "string",
                          "name": "string",
                          "email": "exemple@email.com",
                          "phone": "00 000 000",
                          "role": "customer",
                          "avatar": "string",
                          "statusAccount": true,
                          "adress": "string"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Users not found!"
          },
          "500": {
            "description": "Internal Server Error!"
          }
        }
      }
    },
...
    "/users/confirmAccount/{id}": {
      "patch": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "users"
        ],
        "summary": "Confirm account Restaurant & Deliveryman",
        "description": "Confirm account Restaurant & Deliveryman.",
        "operationId": "confirmAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation Account Confirmed"
          },
          "401": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Access denied!"
          },
          "404": {
            "description": "User not found !"
          },
          "409": {
            "description": "Bad request! Your account is already verified!"
          },
          "500": {
            "description": "Internal Server Error!"
          }
        }
      }
    },
    "/users/request": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "users"
        ],
        "summary": "Get the list of all users unvalidated",
        "description": "Get the list of all users unvalidated.",
        "operationId": "listUsersUnvalidated",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": true,
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successfuly operation ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 1
                    },
                    "totalUsers": {
                      "type": "integer",
                      "example": 1
                    },
                    "users": {
                      "type": "array",
                      "example": [
                        {
                          "_id": "string",
                          "name": "string",
                          "email": "exemple@email.com",
                          "phone": "00 000 000",
                          "role": "customer",
                          "avatar": "string",
                          "statusAccount": false,
                          "adress": "string"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "There are no unvalidated users!"
          },
          "404": {
            "description": "Users not found!"
          },
          "500": {
            "description": "Internal Server Error!"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "required": [
          "name",
          "email",
          "password",
          "role",
          "phone",
          "adress",
          "avatar",
          "statusAccount"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "exemple"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "exemple@email.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "*******"
          },
          "role": {
            "type": "string",
            "example": "Customer",
            "enum": [
              "Restaurant",
              "Deliveryman",
              "Customer"
            ]
          },
          "phone": {
            "type": "integer",
            "example": 12345678
          },
          "adress": {
            "type": "string",
            "example": "example"
          },
          "avatar": {
            "type": "string",
            "example": "example"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    }
  }
}